+++ internal/testui/main.go
@@ -16,7 +16,7 @@ import (
 	"github.com/fatih/color"
 	"github.com/olekukonko/tablewriter"
 
-	"github.com/hashicorp/nomad-pack/sdk/helper"
+	"github.com/hashicorp/nomad-pack/internal/pkg/helper"
 	"github.com/hashicorp/nomad-pack/terminal"
 )
