--- sdk/helper/ptr.go
@@ -1,10 +1,0 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
-package helper
-
-// BoolToPtr returns a pointer to the input boolean value.
-func BoolToPtr(i bool) *bool { return &i }
-
-// StringToPtr returns the pointer to the input string.
-func StringToPtr(str string) *string { return &str }
