+++ sdk/pack/dependency.go
@@ -3,7 +3,7 @@
 
 package pack
 
-import "github.com/hashicorp/nomad-pack/sdk/helper"
+import "github.com/hashicorp/nomad-pack/internal/pkg/helper/pointer"
 
 // Dependency is a single dependency of a pack. A pack can have multiple and
 // each dependency represents an individual pack. A pack can be used as a
@@ -35,7 +35,7 @@ func (d *Dependency) validate() error {
 	}
 
 	if d.Enabled == nil {
-		d.Enabled = helper.BoolToPtr(true)
+		d.Enabled = pointer.Of(true)
 	}
 	return nil
 }
